Fixed a (what I think is) a big usability Bug.
If you look look at the Layer-Tree, how are the images composited?
They should be composited Bottom->Top, but they are not.
This commit fixes that, as it is a 'bug' that is confusing for the user.
(I sat here for 20 damn minutes until I realized that the layers don't
work like they should!)